requirements.txt
discord.py>=2.3.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
matplotlib>=3.8.0
seaborn>=0.13.0
Pillow>=10.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
joblib>=1.3.0